fix: webhook docs links fix: pr review comments
