+++ packages/nc-gui/components/api-client/Headers.vue
@@ -50,6 +50,10 @@ const deleteHeaderRow = (i: number) => vModel.value.splice(i, 1)
 </template>
 
 <style scoped lang="scss">
+.nc-button:not([disabled]) {
+  @apply !text-gray-500;
+}
+
 .ant-input {
   box-shadow: 0px 0px 4px 0px rgba(0, 0, 0, 0.08) !important;
